+++ apps/api/serializers.py
@@ -25,26 +25,15 @@ class UserSerializer(serializers.ModelSerializer):
     """
     Сериалайзер модели User
     """
-    name = serializers.SerializerMethodField(read_only=True)
-    _id = serializers.SerializerMethodField(read_only=True)
     isAdmin = serializers.SerializerMethodField(read_only=True)
 
     class Meta:
         model = User
-        fields = ['id', '_id', 'username', 'email', 'name', 'isAdmin']
-
-    def get__id(self, obj):
-        return obj.id
+        fields = ['id', 'first_name', 'last_name', 'email', 'description', 'phone_number', 'rate', 'job_title',
+                  'isAdmin']
 
     def get_isAdmin(self, obj):
         return obj.is_staff
-
-    def get_name(self, obj):
-        name = obj.first_name
-        if name == '':
-            name = obj.email
-
-        return name
 
 
 class UserSerializerWithToken(UserSerializer):