+++ apps/api/urls/user_urls.py
@@ -7,6 +7,7 @@ urlpatterns = [
     path('login/', views.MyTokenObtainPairView.as_view(),
          name='token_obtain_pair'),
 
-    path('register/', views.register_user, name='register'),
-    path('profile/', views.getUserProfile, name="users-profile"),
+    path('register/', views.register_user),
+    path('profile/', views.getUserProfile),
+    path('user/edit/', views.updateUserProfile),
 ]